Move the configuration for API servers to its own file.
