+++ cmd/notary-signer/config.go
@@ -1,11 +1,17 @@
 package main
 
 import (
+	"crypto/tls"
 	"errors"
 	"fmt"
+	"net"
+	"net/http"
 	"os"
 	"strings"
 	"time"
+
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/dancannon/gorethink"
@@ -13,7 +19,9 @@ import (
 	"github.com/docker/notary"
 	"github.com/docker/notary/cryptoservice"
 	"github.com/docker/notary/passphrase"
+	pb "github.com/docker/notary/proto"
 	"github.com/docker/notary/signer"
+	"github.com/docker/notary/signer/api"
 	"github.com/docker/notary/signer/keydbstore"
 	"github.com/docker/notary/storage/rethinkdb"
 	"github.com/docker/notary/trustmanager"
@@ -107,4 +115,40 @@ func getDefaultAlias(configuration *viper.Viper) (string, error) {
 	}
 	logrus.Debug("Default Alias: ", defaultAlias)
 	return defaultAlias, nil
+}
+
+// set up the GRPC server
+func setupGRPCServer(grpcAddr string, tlsConfig *tls.Config,
+	cryptoServices signer.CryptoServiceIndex) (*grpc.Server, net.Listener, error) {
+
+	//RPC server setup
+	kms := &api.KeyManagementServer{CryptoServices: cryptoServices,
+		HealthChecker: health.CheckStatus}
+	ss := &api.SignerServer{CryptoServices: cryptoServices,
+		HealthChecker: health.CheckStatus}
+
+	lis, err := net.Listen("tcp", grpcAddr)
+	if err != nil {
+		return nil, nil, fmt.Errorf("grpc server failed to listen on %s: %v",
+			grpcAddr, err)
+	}
+
+	creds := credentials.NewTLS(tlsConfig)
+	opts := []grpc.ServerOption{grpc.Creds(creds)}
+	grpcServer := grpc.NewServer(opts...)
+
+	pb.RegisterKeyManagementServer(grpcServer, kms)
+	pb.RegisterSignerServer(grpcServer, ss)
+
+	return grpcServer, lis, nil
+}
+
+func setupHTTPServer(httpAddr string, tlsConfig *tls.Config,
+	cryptoServices signer.CryptoServiceIndex) *http.Server {
+
+	return &http.Server{
+		Addr:      httpAddr,
+		Handler:   api.Handlers(cryptoServices),
+		TLSConfig: tlsConfig,
+	}
 }
